+++ server/modules/index.js
@@ -64,11 +64,116 @@ module.exports = [
         dependencies: ['instantiation.consoleLogger'],
         module: require(__dirname + '/instantiation/instance-classes')
     },
+
+    {
+        name: 'instantiation.job.baseClasses',
+        dependencies: [],
+        module: require(__dirname + '/instantiation/job/base-classes')
+    },
+    {
+        name: 'instantiation.job.resolveReference',
+        dependencies: [
+            'instantiation.job.baseClasses',
+            'github.apiClient'
+        ],
+        module: require(__dirname + '/instantiation/job/resolve-reference')
+    },
+    {
+        name: 'instantiation.job.createDirectory',
+        dependencies: [
+            'instantiation.job.baseClasses'
+        ],
+        module: require(__dirname + '/instantiation/job/create-directory')
+    },
+    {
+        name: 'instantiation.job.downloadArchive',
+        dependencies: [
+            'config',
+            'instantiation.job.baseClasses'
+        ],
+        module: require(__dirname + '/instantiation/job/download-archive')
+    },
+    {
+        name: 'instantiation.job.extractArchive',
+        dependencies: [
+            'instantiation.job.baseClasses'
+        ],
+        module: require(__dirname + '/instantiation/job/extract-archive')
+    },
+    {
+        name: 'instantiation.job.providePort',
+        dependencies: [
+            'instantiation.job.baseClasses',
+            'instantiation.portProvider',
+            'mongodb.repository.buildInstance'
+        ],
+        module: require(__dirname + '/instantiation/job/provide-port')
+    },
+    {
+        name: 'instantiation.job.copyBeforeBuildTask',
+        dependencies: [
+            'instantiation.job.baseClasses'
+        ],
+        module: require(__dirname + '/instantiation/job/copy-before-build-task')
+    },
+    {
+        name: 'instantiation.job.interpolateBeforeBuildTask',
+        dependencies: [
+            'instantiation.job.baseClasses',
+            'instantiation.interpolationHelper'
+        ],
+        module: require(__dirname + '/instantiation/job/interpolate-before-build-task')
+    },
+    {
+        name: 'instantiation.job.prepareEnvironmentalVariables',
+        dependencies: [
+            'instantiation.job.baseClasses',
+            'instantiation.interpolationHelper',
+            'mongodb.repository.buildInstance'
+        ],
+        module: require(__dirname + '/instantiation/job/prepare-environmental-variables')
+    },
+    {
+        name: 'instantiation.job.prepareSummaryItems',
+        dependencies: [
+            'instantiation.job.baseClasses',
+            'instantiation.interpolationHelper',
+            'mongodb.repository.buildInstance'
+        ],
+        module: require(__dirname + '/instantiation/job/prepare-summary-items')
+    },
+    {
+        name: 'instantiation.job.runDockerCompose',
+        dependencies: [
+            'instantiation.job.baseClasses'
+        ],
+        module: require(__dirname + '/instantiation/job/run-docker-compose')
+    },
+
+    {
+        name: 'instantiation.job.executor',
+        dependencies: [],
+        module: require(__dirname + '/instantiation/job/executor')
+    },
+
     {
         name: 'instantiation.jobs',
-        dependencies: ['config', 'instantiation.portProvider', 'instantiation.interpolationHelper', 'mongodb.repository.buildInstance', 'github.apiClient'],
+        dependencies: [
+            'instantiation.job.resolveReference',
+            'instantiation.job.createDirectory',
+            'instantiation.job.downloadArchive',
+            'instantiation.job.extractArchive',
+            'instantiation.job.providePort',
+            'instantiation.job.copyBeforeBuildTask',
+            'instantiation.job.interpolateBeforeBuildTask',
+            'instantiation.job.prepareEnvironmentalVariables',
+            'instantiation.job.prepareSummaryItems',
+            'instantiation.job.runDockerCompose',
+            'instantiation.job.executor'
+        ],
         module: require(__dirname + '/instantiation/jobs')
     },
+
     {
         name: 'instantiation.sourceCodeFetcher',
         dependencies: ['config', 'instantiation.instanceClasses', 'instantiation.jobs'],
