+++ server/modules/instantiation/job/interpolate-before-build-task.js
@@ -1,0 +1,43 @@
+var path = require('path');
+var Promise = require('bluebird');
+
+module.exports = function (
+    baseClasses,
+    interpolationHelper
+) {
+
+    var {ComponentInstanceJob, JobExecutor} = baseClasses;
+
+    class InterpolateBeforeBuildTaskJob extends ComponentInstanceJob {
+        constructor(componentInstance, relativePath) {
+            super(componentInstance);
+            this.relativePath = relativePath;
+        }
+    }
+
+    class InterpolateBeforeBuildTaskJobExecutor extends JobExecutor {
+        supports(job) {
+            return (job instanceof InterpolateBeforeBuildTaskJob);
+        }
+
+        execute(job) {
+            var { componentInstance } = job;
+
+            return new Promise((resolve) => {
+                var fullPath = path.join(componentInstance.fullPath, job.relativePath);
+                var { featVariables, externalPorts } = componentInstance.buildInstance;
+
+                componentInstance.log(`Interpolating Feat variables in ${job.relativePath}.`);
+                interpolationHelper.interpolateFile(fullPath, featVariables, externalPorts);
+
+                resolve();
+            });
+        }
+    }
+
+    return {
+        InterpolateBeforeBuildTaskJob,
+        InterpolateBeforeBuildTaskJobExecutor
+    };
+
+};
